src/drone_agent/requirements.txt
mavsdk>=1.4.0
jpl-rosa>=1.0.7
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
rich>=13.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0